+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Improvements
 
+-   Implement customTimeout for resource deletion. (https://github.com/pulumi/pulumi-kubernetes/pull/802).
 -   Warn for invalid usage of Helm repo parameter. (https://github.com/pulumi/pulumi-kubernetes/pull/805).
 
 ## 1.0.1 (September 11, 2019)